Move tag modals to their own file
